Refactored running notes and links, enabled them for Worksets, Flowcells.
