+++ run_dir/design/link_tab.html
@@ -1,0 +1,45 @@
+            <div class="tab-pane fade" id="tab_links_content">
+                <div class="row" style="margin-bottom:30px;">
+                    <div id="existing_links"></div>
+                    </div>
+
+                    <form role="form" id="link_form" class="form-horizontal">
+                        <div id="new_link_alert"></div>
+                        <div class="form-group">
+                          <label for="new_link_type" class="col-sm-2 control-label">Type</label>
+                          <div class="col-sm-9">
+                            <select class="form-control" required id="new_link_type">
+                              <option value="">[ choose link type ]</option>
+                              <option>Deviation</option>
+                              <option>Other</option>
+                            </select>
+                          </div>
+                        </div>
+                        <div class="form-group">
+                          <label for="new_link_title" class="col-sm-2 control-label">Title</label>
+                          <div class="col-sm-9">
+                            <input type="text" required class="form-control span8" id="new_link_title" placeholder="14_01 example"/>
+                          </div>
+                        </div>
+                        <div class="form-group">
+                          <label for="new_link_url" class="col-sm-2 control-label">URL</label>
+                          <div class="col-sm-9">
+                            <input type="url" class="form-control span8" id="new_link_url" placeholder="https://docs.google.com/example"/>
+                          </div>
+                        </div>
+                        <div class="form-group">
+                          <label for="new_link_desc" class="col-sm-2 control-label">Description</label>
+                          <div class="col-sm-9">
+                            <textarea rows="4" class="form-control span8" id="new_link_desc" placeholder="Description text"></textarea>
+                          </div>
+                        </div>
+                        <div class="form-group">
+                          <label for="save_link_button" class="col-sm-2 control-label"></label>
+                          <div class="col-sm-9">
+                            <button type="submit" class="btn btn-primary" id="save_link_button">Add link</button>
+                          </div>
+                        </div>
+                      </form>
+                    </div>
+                </div>
+            </div>
