+++ run_dir/design/running_notes_tab.html
@@ -1,0 +1,33 @@
+    <!-- RUNNING NOTES TAB -->
+    <div class="tab-pane fade" id="tab_running_notes_content">
+        <!-- add running note -->
+        <form action="" method="POST" id="running_notes_form" role="form">
+            <div class="panel panel-info">
+                <div class="panel-heading">Add New Running Note</div>
+                <div class="panel-body">
+                    <div class="row">
+                        <div class="col-md-6">
+                            <h4>Write Running Note</h4>
+                            <textarea rows="5" class="form-control" id="new_note_text" style="height:97px;"></textarea>
+                        </div>
+                        <div class="col-md-6">
+                            <h4>Preview</h4>
+                            <div class="panel panel-default" id="running_note_preview_panel">
+                                <div class="panel-heading"><a href="#">{{ user }}</a> - <span class="todays_date">Date</span></div>
+                                <div class="panel-body" id="running_note_preview_body">
+                                    <p class="text-muted"><em>Nothing to preview..</em></p>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="text-right" style="margin-top:15px;">
+                        <button type="button" class="btn btn-link" data-toggle="modal" data-target="#markdown_help">Markdown Help</a>
+                        <button type="submit" class="btn btn-primary" id="save_note_button">Submit Running Note</button>
+                    </div>
+                </div>
+            </div>
+        </form>
+
+        <!-- display running notes -->
+        <div class="panel-group" id="running_notes_panels"></div>
+    </div>
